requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
binaryornot==0.4.4
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
crispy-bootstrap5==2024.2
Django==4.2.11
django-allauth==0.62.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.5.0
django-redis==5.4.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
hiredis==2.3.2
idna==3.7
inflection==0.5.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
Pygments==2.17.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.4
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
